CI + Updates for OS X 10.10.

+++ Device Monitor/Classes/Controllers/FileSystem/List/DMFileSystemPartitionTableViewController+UITableViewDataSource.m
@@ -70,7 +70,7 @@
     
     @try
     {
-        cell.partition = [ _partitions objectAtIndex: indexPath.row ];
+        cell.partition = [ _partitions objectAtIndex: ( NSUInteger )( indexPath.row ) ];
     }
     @catch ( NSException * e )
     {
